// template/apps/api/src/users/entities/user.entity.ts
import { ApiProperty } from '@nestjs/swagger';
import { User } from '@workspace/database';
import { Exclude } from 'class-transformer';
import {
  TodoEntity,
  type TodoEntity as TodoEntityAsType,
} from 'src/todos/entities/todo.entity';

export class UserEntity implements User {
  /*
  this code was not generated by prisma-generator-nestjs-dto
  but manually added to demonstrate usage of decorators along with @nestjs/swagger Exluce/Expose Properties
  the reason to implement an entity like this, is to be able to exlude certain properties from the response 
  and to document it accordingly with the @nestjs/swagger @ApiReponse decorators
  e.g. password fields, createdAt, updatedAt, etc.
  an example is found in this prisma-nestjs documentation: https://www.prisma.io/blog/nestjs-prisma-relational-data-7D056s1kOabc#generate-new-users-rest-resource
  */
  constructor(partial: Partial<UserEntity>) {
    Object.assign(this, partial);
  }

  @ApiProperty({
    type: 'string',
  })
  id: string;
  @ApiProperty({
    description: 'Firstname of the user',
    minimum: 5,
    type: 'string',
    nullable: true,
  })
  firstName: string | null;
  @ApiProperty({
    description: 'Firstname of the user',
    minimum: 5,
    type: 'string',
    nullable: true,
  })
  lastName: string;
  @ApiProperty({
    description: 'Email of the user',
    type: 'string',
    required: true,
    nullable: true,
  })
  email!: string;

  @ApiProperty({
    description: 'Hashed password of the user',
    minimum: 10,
    type: 'string',
  })
  @Exclude()
  password: string;

  @ApiProperty({
    type: () => TodoEntity,
    isArray: true,
    required: false,
  })
  todos?: TodoEntityAsType[];
}